Fix sorting by modified date. #1467
Disable sorting by score in iBus because from InGrid 5.0.0 onwards
all results come from one iPlug (in this case index) and it is
considered that th iPlug cares for sorting itself.

In case of sorting by modification date the scores of the results do not
reflect the sort order and sorting in iBus would destroy the sort
order my modification date.